Add authorization check for deleting admin users

+++ src/app/Components/users/users.component.ts
@@ -15,6 +15,7 @@ import {
   Validators,
 } from '@angular/forms';
 import { DropdownModule } from 'primeng/dropdown';
+import { AuthService } from '../../Services/auth.service';
 
 @Component({
   selector: 'app-users',
@@ -39,7 +40,8 @@ export class UsersComponent implements OnInit {
     private _UserService: UserService,
     private _MessageService: MessageService,
     private _Router: Router,
-    private _ActivatedRoute: ActivatedRoute
+    private _ActivatedRoute: ActivatedRoute,
+    private _AuthService: AuthService
   ) {}
 
   users: User[] = [];
@@ -172,24 +174,32 @@ export class UsersComponent implements OnInit {
   }
 
   deleteUser(user: User) {
-    this._UserService.deleteUser(user._id).subscribe({
-      next: (res) => {
-        this.users = this.users.filter((u) => u._id !== user._id);
-        this._MessageService.add({
-          severity: 'success',
-          summary: 'Success',
-          detail: res.message,
-        });
-      },
-      error: (err) => {
-        console.error(err);
-        this._MessageService.add({
-          severity: 'error',
-          summary: 'Error',
-          detail: err.error.message,
-        });
-      },
-    });
+    if (this._AuthService.currentUser.value?.rule === 'admin') {
+      this._MessageService.add({
+        severity: 'error',
+        summary: 'Error',
+        detail: 'You are not allowed to delete a admin user',
+      });
+    } else {
+      this._UserService.deleteUser(user._id).subscribe({
+        next: (res) => {
+          this.users = this.users.filter((u) => u._id !== user._id);
+          this._MessageService.add({
+            severity: 'success',
+            summary: 'Success',
+            detail: res.message,
+          });
+        },
+        error: (err) => {
+          console.error(err);
+          this._MessageService.add({
+            severity: 'error',
+            summary: 'Error',
+            detail: err.error.message,
+          });
+        },
+      });
+    }
   }
 }
